Renamed field_group to fieldGroup to be consistent with other files (#1417)
