+++ imports/plugins/core/ui/client/components/index.js
@@ -8,4 +8,4 @@ export { default as FlatButton } from "./button/flatButton";
 export { default as IconButton } from "./button/iconButton";
 export { default as EditButton } from "./button/editButton";
 export { default as Loading } from "./loading/loading";
-export { default as FieldGroup } from "./forms/field_group";
+export { default as FieldGroup } from "./forms/fieldGroup";
